requirements.txt
httpx==0.24.1
httpcore==0.17.3
openai==1.12.0
pinecone-client==3.1.0
neo4j==5.16.0
python-dotenv==1.0.0
tiktoken==0.5.2
numpy==1.26.4
pandas==2.2.0
aiohttp==3.9.3
tenacity==8.2.3
tqdm==4.66.2
jupyter==1.0.0
ipython==8.21.0
pyvis==0.3.2
networkx==3.2.1
google-generativeai==0.3.2
black==24.1.0
flake8>=6.0.0